rewrote the main loop to use poll
We can handle up to MAX_USERS (64 by default) concurrently.

Now, given that we don’t support CGI, it’s not a big deal.  Gemini
requests are small (up to 1024 bytes), and also the replies from the
server are small (one line plus the document — if any), all over TLS
obviously. (but even there, it’s lighter than HTTP because we don’t
need to send the whole chain for the certificate — see TOFU).

Given all the above, this doesn’t really improve the performance in
the real world, but it’s nice to have.  The main use case for this is
to disallow slow clients to stop fast clients.