put back the missing lseek, and change variables name
(I know, changing variables names AND introducing changes is better
done in more commits, but…)

Added back an lseek that was missing.  If TLS_WANT_POLL{IN,OUT}, we
need to re-send that block, but we need also to rewind the file, in
order to read(2) that chunk again.  This doesn’t solve the corruption
in transferring big files, but reduces them.  I still haven’t tracked
down the corruption :/