add an implicit fastcgi parameter: GEMINI_SEARCH_STRING
backport of 97b306cbee6d105885a761e04274f661a0ec3757

it’s the QUERY_STRING decoded if it’s a search-string (i.e. not a
key-value pair.)  It’s useful for scripts to avoid percent-decoding
the querystring in the most common case of a query, because in Gemini
querystrings key-value paired are not common.

Idea from a discussion with Allen Sobot.