proxy-protocol: accept cross-family proxying
Due to a strict interpretation of the spec if "TCP4" is used we
expect two ipv4 addresses (and similar for "TCP6" and ipv6 addresses).
However, the family specified in the proxy header matters only for
the first address (the source), not the destination!  After all,
it's not strange to proxy from/to ipv4 and ipv6.

Use getaddrinfo(NI_NUMERICHOST) to parse the IP addresses since
inet_pton() is too strict.